Move general history schemas to pages
