+++ src/applications/pensions/config/form.js
@@ -5,7 +5,6 @@ import moment from 'moment';
 import { createSelector } from 'reselect';
 
 import fullSchemaPensions from 'vets-json-schema/dist/21P-527EZ-schema.json';
-import { isFullDate } from 'platform/forms/validations';
 import { externalServices } from 'platform/monitoring/DowntimeNotification';
 import FormFooter from 'platform/forms/components/FormFooter';
 import environment from 'platform/utilities/environment';
@@ -15,7 +14,6 @@ import * as address from 'platform/forms/definitions/address';
 import bankAccountUI from 'platform/forms/definitions/bankAccount';
 import { VA_FORM_IDS } from 'platform/forms/constants';
 
-import FullNameField from 'platform/forms-system/src/js/fields/FullNameField';
 import currentOrPastDateUI from 'platform/forms-system/src/js/definitions/currentOrPastDate';
 import phoneUI from 'platform/forms-system/src/js/definitions/phone';
 import fullNameUI from 'platform/forms/definitions/fullName';
@@ -66,6 +64,7 @@ import { additionalSourcesSchema } from '../definitions/additionalSources';
 import otherExpensesUI from '../definitions/otherExpenses';
 import applicantInformation from '../pages/applicantInformation';
 import servicePeriods from '../pages/servicePeriods';
+import generalHistory from '../pages/generalHistory';
 
 import {
   validateAfterMarriageDate,
@@ -79,10 +78,7 @@ const {
   nationalGuardActivation,
   nationalGuard,
   disabilities,
-  previousNames,
-  combatSince911,
   jobs,
-  placeOfSeparation,
   powDateRange,
   severancePay,
   spouseDateOfBirth,
@@ -260,63 +256,8 @@ const formConfig = {
         general: {
           path: 'military/general',
           title: 'General history',
-          uiSchema: {
-            'view:serveUnderOtherNames': {
-              'ui:title': 'Did you serve under another name?',
-              'ui:widget': 'yesNo',
-            },
-            previousNames: {
-              'ui:options': {
-                itemName: 'Name',
-                expandUnder: 'view:serveUnderOtherNames',
-                viewField: FullNameField,
-                reviewTitle: 'Previous names',
-              },
-              items: fullNameUI,
-            },
-            placeOfSeparation: {
-              'ui:title':
-                'Place of last or anticipated separation (city and state or foreign country)',
-            },
-            combatSince911: (() => {
-              const rangeExcludes911 = createSelector(
-                form => form.servicePeriods,
-                periods =>
-                  (periods || []).every(
-                    period =>
-                      !period.activeServiceDateRange ||
-                      !isFullDate(period.activeServiceDateRange.to) ||
-                      !moment('2001-09-11').isBefore(
-                        period.activeServiceDateRange.to,
-                      ),
-                  ),
-              );
-
-              return {
-                'ui:title': 'Did you serve in a combat zone after 9/11/2001?',
-                'ui:widget': 'yesNo',
-                'ui:required': formData => !rangeExcludes911(formData),
-                'ui:options': {
-                  hideIf: rangeExcludes911,
-                },
-              };
-            })(),
-          },
-          schema: {
-            type: 'object',
-            required: ['view:serveUnderOtherNames'],
-            properties: {
-              'view:serveUnderOtherNames': {
-                type: 'boolean',
-              },
-              previousNames: {
-                ...previousNames,
-                minItems: 1,
-              },
-              placeOfSeparation,
-              combatSince911,
-            },
-          },
+          uiSchema: generalHistory.uiSchema,
+          schema: generalHistory.schema,
         },
         reserveAndNationalGuard: {
           path: 'military/reserve-national-guard',
